restructure commit
